requirements.txt
alembic==1.6.2
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mysql-connector-python==8.0.24
mysqlclient==2.0.3
passlib==1.7.4
protobuf==3.15.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.4.7
SQLAlchemy-Utils==0.36.8
sqlparse==0.4.1
Werkzeug==1.0.1
WTForms==2.3.3
